langchain/langchain-course/section-3-search-agents/pyproject.toml
[project]
name = "section-3-search-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "isort>=6.0.1",
    "langchain-openai>=0.3.33",
    "langchain-tavily>=0.2.11",
    "python-dotenv>=1.1.1",
]
